Set versions

+++ tools/update-components.sh
@@ -19,38 +19,8 @@ if [ ! -d "$AR_COMPS/arduino" ]; then
 	git clone $AR_REPO_URL "$AR_COMPS/arduino"
 fi
 
-if [ -z $AR_BRANCH ]; then
-	if [ -z $GITHUB_HEAD_REF ]; then
-		current_branch=`git branch --show-current`
-	else
-		current_branch="$GITHUB_HEAD_REF"
-	fi
-	echo "Current Branch: $current_branch"
-	if [[ "$current_branch" != "master" && `git_branch_exists "$AR_COMPS/arduino" "$current_branch"` == "1" ]]; then
-		export AR_BRANCH="$current_branch"
-	else
-		if [ -z "$IDF_COMMIT" ]; then #commit was not specified at build time
-			AR_BRANCH_NAME="idf-$IDF_BRANCH"
-		else
-			AR_BRANCH_NAME="idf-$IDF_COMMIT"
-		fi
-		has_ar_branch=`git_branch_exists "$AR_COMPS/arduino" "$AR_BRANCH_NAME"`
-		if [ "$has_ar_branch" == "1" ]; then
-			export AR_BRANCH="$AR_BRANCH_NAME"
-		else
-			has_ar_branch=`git_branch_exists "$AR_COMPS/arduino" "$AR_PR_TARGET_BRANCH"`
-			if [ "$has_ar_branch" == "1" ]; then
-				export AR_BRANCH="$AR_PR_TARGET_BRANCH"
-			fi
-		fi
-	fi
-fi
-
-if [ "$AR_BRANCH" ]; then
-	git -C "$AR_COMPS/arduino" checkout "$AR_BRANCH" && \
-	git -C "$AR_COMPS/arduino" fetch && \
-	git -C "$AR_COMPS/arduino" pull --ff-only
-fi
+git -C "$AR_COMPS/arduino" fetch && \
+git -C "$AR_COMPS/arduino" checkout 2.0.6
 if [ $? -ne 0 ]; then exit 1; fi
 
 #
@@ -60,13 +30,14 @@ echo "Updating ESP32 Camera..."
 if [ ! -d "$AR_COMPS/esp32-camera" ]; then
 	git clone $CAMERA_REPO_URL "$AR_COMPS/esp32-camera"
 else
-	git -C "$AR_COMPS/esp32-camera" fetch && \
-	git -C "$AR_COMPS/esp32-camera" pull --ff-only
+	git -C "$AR_COMPS/esp32-camera" fetch
 fi
+if [ $? -ne 0 ]; then exit 1; fi
 #this is a temp measure to fix build issue
 # if [ -f "$AR_COMPS/esp32-camera/idf_component.yml" ]; then
 # 	rm -rf "$AR_COMPS/esp32-camera/idf_component.yml"
 # fi
+git -C "$AR_COMPS/esp32-camera" checkout "$(grep -e ^esp32-camera: $AR_COMPS/arduino/tools/sdk/versions.txt | cut -d ' ' -f 3)"
 if [ $? -ne 0 ]; then exit 1; fi
 
 #
@@ -76,9 +47,9 @@ echo "Updating ESP-DL..."
 if [ ! -d "$AR_COMPS/esp-dl" ]; then
 	git clone $DL_REPO_URL "$AR_COMPS/esp-dl"
 else
-	git -C "$AR_COMPS/esp-dl" fetch && \
-	git -C "$AR_COMPS/esp-dl" pull --ff-only
+	git -C "$AR_COMPS/esp-dl" fetch
 fi
+git -C "$AR_COMPS/esp-dl" checkout "$(grep -e ^esp-dl: $AR_COMPS/arduino/tools/sdk/versions.txt | cut -d ' ' -f 3)"
 if [ $? -ne 0 ]; then exit 1; fi
 
 #
@@ -88,9 +59,10 @@ echo "Updating ESP-SR..."
 if [ ! -d "$AR_COMPS/esp-sr" ]; then
 	git clone $SR_REPO_URL "$AR_COMPS/esp-sr"
 else
-	git -C "$AR_COMPS/esp-sr" fetch && \
-	git -C "$AR_COMPS/esp-sr" pull --ff-only
+	git -C "$AR_COMPS/esp-sr" fetch
 fi
+git -C "$AR_COMPS/esp-dl" checkout "$(grep -e ^esp-dl: $AR_COMPS/arduino/tools/sdk/versions.txt | cut -d ' ' -f 3)"
+if [ $? -ne 0 ]; then exit 1; fi
 #this is a temp measure to fix build issue
 if [ -f "$AR_COMPS/esp-sr/idf_component.yml" ]; then
 	rm -rf "$AR_COMPS/esp-sr/idf_component.yml"
@@ -102,13 +74,13 @@ if [ $? -ne 0 ]; then exit 1; fi
 #
 echo "Updating ESP-LITTLEFS..."
 if [ ! -d "$AR_COMPS/esp_littlefs" ]; then
-	git clone $LITTLEFS_REPO_URL "$AR_COMPS/esp_littlefs" && \
-    git -C "$AR_COMPS/esp_littlefs" submodule update --init --recursive
+	git clone $LITTLEFS_REPO_URL "$AR_COMPS/esp_littlefs"
 else
-	git -C "$AR_COMPS/esp_littlefs" fetch && \
-	git -C "$AR_COMPS/esp_littlefs" pull --ff-only && \
-    git -C "$AR_COMPS/esp_littlefs" submodule update --init --recursive
+	git -C "$AR_COMPS/esp_littlefs" fetch
 fi
+if [ $? -ne 0 ]; then exit 1; fi
+git -C "$AR_COMPS/esp_littlefs" checkout "$(grep -e ^esp_littlefs: $AR_COMPS/arduino/tools/sdk/versions.txt | cut -d ' ' -f 3)" && \
+git -C "$AR_COMPS/esp_littlefs" submodule update --init --recursive
 if [ $? -ne 0 ]; then exit 1; fi
 
 #
@@ -116,13 +88,13 @@ if [ $? -ne 0 ]; then exit 1; fi
 #
 echo "Updating ESP-RainMaker..."
 if [ ! -d "$AR_COMPS/esp-rainmaker" ]; then
-    git clone $RMAKER_REPO_URL "$AR_COMPS/esp-rainmaker" && \
-    git -C "$AR_COMPS/esp-rainmaker" submodule update --init --recursive
+    git clone $RMAKER_REPO_URL "$AR_COMPS/esp-rainmaker"
 else
-	git -C "$AR_COMPS/esp-rainmaker" fetch && \
-	git -C "$AR_COMPS/esp-rainmaker" pull --ff-only && \
-    git -C "$AR_COMPS/esp-rainmaker" submodule update --init --recursive
+	git -C "$AR_COMPS/esp-rainmaker" fetch
 fi
+if [ $? -ne 0 ]; then exit 1; fi
+git -C "$AR_COMPS/esp-rainmaker" checkout "$(grep -e ^esp-rainmaker: $AR_COMPS/arduino/tools/sdk/versions.txt | cut -d ' ' -f 3)" && \
+git -C "$AR_COMPS/esp-rainmaker" submodule update --init --recursive
 if [ $? -ne 0 ]; then exit 1; fi
 
 #
@@ -130,13 +102,13 @@ if [ $? -ne 0 ]; then exit 1; fi
 #
 echo "Updating ESP-Insights..."
 if [ ! -d "$AR_COMPS/esp-insights" ]; then
-    git clone $INSIGHTS_REPO_URL "$AR_COMPS/esp-insights" && \
-    git -C "$AR_COMPS/esp-insights" submodule update --init --recursive
+    git clone $INSIGHTS_REPO_URL "$AR_COMPS/esp-insights"
 else
-	git -C "$AR_COMPS/esp-insights" fetch && \
-	git -C "$AR_COMPS/esp-insights" pull --ff-only && \
-    git -C "$AR_COMPS/esp-insights" submodule update --init --recursive
+	git -C "$AR_COMPS/esp-insights" fetch
 fi
+if [ $? -ne 0 ]; then exit 1; fi
+git -C "$AR_COMPS/esp-insights" checkout "$(grep -e ^esp-insights: $AR_COMPS/arduino/tools/sdk/versions.txt | cut -d ' ' -f 3)" && \
+git -C "$AR_COMPS/esp-insights" submodule update --init --recursive
 if [ $? -ne 0 ]; then exit 1; fi
 
 #
@@ -146,9 +118,10 @@ echo "Updating ESP-DSP..."
 if [ ! -d "$AR_COMPS/esp-dsp" ]; then
 	git clone $DSP_REPO_URL "$AR_COMPS/esp-dsp"
 else
-	git -C "$AR_COMPS/esp-dsp" fetch && \
-	git -C "$AR_COMPS/esp-dsp" pull --ff-only
+	git -C "$AR_COMPS/esp-dsp" fetch
 fi
+if [ $? -ne 0 ]; then exit 1; fi
+git -C "$AR_COMPS/esp-dsp" checkout "$(grep -e ^esp-dsp: $AR_COMPS/arduino/tools/sdk/versions.txt | cut -d ' ' -f 3)"
 if [ $? -ne 0 ]; then exit 1; fi
 
 #
@@ -158,8 +131,8 @@ echo "Updating TinyUSB..."
 if [ ! -d "$AR_COMPS/arduino_tinyusb/tinyusb" ]; then
 	git clone $TINYUSB_REPO_URL "$AR_COMPS/arduino_tinyusb/tinyusb"
 else
-	git -C "$AR_COMPS/arduino_tinyusb/tinyusb" fetch && \
-	git -C "$AR_COMPS/arduino_tinyusb/tinyusb" pull --ff-only
+	git -C "$AR_COMPS/arduino_tinyusb/tinyusb" fetch
 fi
 if [ $? -ne 0 ]; then exit 1; fi
-
+git -C "$AR_COMPS/arduino_tinyusb/tinyusb" checkout "$(grep -e ^tinyusb: $AR_COMPS/arduino/tools/sdk/versions.txt | cut -d ' ' -f 3)"
+if [ $? -ne 0 ]; then exit 1; fi
